Don't allocate dynamic containers by default
The code that creates the container might not know the required capacity,
but it might still be known before the container is filled. Default
allocating too little memory would force two allocations in that case.
Let's give power to the users since having a default size seems like a
pessimization anyway.

Don't apply this to String because it has to be a valid c-string while
unallocated. c_str() could give out a pointer to known zeroed mem (like
m_size) when m_data == nullptr, but that seems dirty until an actual
use case arises.